requirements.txt
requests>=2.28.0
colorama>=0.4.0
shodan>=1.28.0
phonenumbers>=8.13.0
python-whois>=0.8.0
dnspython>=2.4.0
socialscan>=1.0.0
